This commit rebalances all ingredient and recipe costs and fixes hamburger bug
This commit rebalances all ingredient and recipe costs to create a more forgiving early game and a smoother progression curve.

It also fixes a game-breaking bug where the Burger recipe was impossible to unlock in Aspen Peak by changing the competition requirement to the Reuben sandwich. City preferences have been updated to create a more logical travel path for the player.